refactor(deeptrace): split detectors into modules
